v3/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-jose
passlib
boto3
sentence-transformers
faiss-cpu
pymupdf
python-docx
python-pptx
pandas
openpyxl
pytest
httpx
torch
torchaudio
psutil
torchvision
